requirements.txt
anndata==0.10.9
antiorm==1.2.1
array_api_compat==1.9
biopython==1.85
biothings-client==0.3.1
boto3==1.37.23
botocore==1.37.23
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
db==0.1.1
exceptiongroup==1.2.2
GEOparse==2.0.4
h5py==3.12.1
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
lxml==5.4.0
markdown-it-py==3.0.0
mdurl==0.1.2
mygene==3.2.2
natsort==8.4.0
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pluggy==1.5.0
psycopg2-binary==2.9.10
Pygments==2.19.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
s3transfer==0.11.4
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
SQLAlchemy==2.0.40
tdqm==0.0.1
tqdm==4.66.5
typer==0.15.2
typing_extensions==4.13.0
tzdata==2024.2
urllib3==2.2.3
xmltodict==0.14.2
